requirements.txt
httpx==0.27.0
websockets==12.0
finnhub-python==2.4.20
ollama==0.3.3
runpod==1.7.3
python-telegram-bot==21.5
resend==2.3.0
supabase==2.7.4
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.36
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
